start again - work on sep portions
