+++ students/csimmons/lesson03/mailroom.py
@@ -21,81 +21,6 @@ userprompt = '\n'.join(('What would you like to do',
           'Please choose from the options below:',
           '1 - Send a Thank You letter',
           '2 - Create a report',
-          '3 - Add data',
+          '3 - Add new donation data',
           '4 - Quit',
           '>>> '))
-
-def exit_option():
-    print('Thank you and goodbye!')
-    sys.exit()
-
-def option_one():
-    print('1 was selected')
-    sys.exit()
-
-def option_two():
-    print('2 was selected')
-    sys.exit()
-
-def option_three():
-    print('3 was selected')
-    sys.exit()
-
-def option_four():
-    print('4 was selected')
-    sys.exit()
-
-def unpack(seq):
-    new_donors =[]
-    sum_this =[]
-    for i in range(len(seq)):
-        new_donors.append(seq[i][0])
-        print(seq[i][0])
-        sum_this.append(seq[i][1])
-        print(seq[i][1])
-    
-
-    
-    print(new_donors)
-    print(sum_this)
-
-unpack(donorlist)
-#print(mailroom_data1)
-
-def mailroom(seq):
-    new_donors =[]
-    cash_only = []
-    for i in range(len(seq)):
-        print((seq[i][0]))
-        print((seq[i][1:]))
-        new_donors.append(seq[i][0])
-        cash_only.append(seq[i][1:])
-    print(new_donors)
-    print(cash_only)
-    print(cash_only[1])
-
-def main():
-    response = input(userprompt)
-    while(True):
-        if response == '1':
-            option_one()
-        elif response == '2':
-            option_two()
-        elif response == '3':
-            option_three()
-        elif response == '4':
-            exit_option()
-        else:
-            print('Invalid')
-            break
-
-#main()
-#mailroom(donorlist2)
-
-
-
-
-'''
-if _name_ == '_main_':
-    main()
-'''
